(1) Remove remove-hidden-jobs-manager import (moved to settings.js as display-settings.js)
(2) Update references
