--- extension/chrome/runtime/popup/classes/job-board-popup.js
@@ -1,20 +1,0 @@
-import { RemoveHiddenJobsManager } from "./remove-hidden-jobs-manager.js";
-import { HiddenJobsListManager } from "./hidden-jobs-list-manager.js";
-import { UnblockAllJobsManager } from "./unblock-all-jobs-manager.js";
-
-class JobBoardPopup {
-  static started = false;
-  static async start(jobBoard) {
-    document.documentElement.setAttribute("data-job-board-id", jobBoard.id);
-
-    if (this.started) return;
-    this.started = true;
-
-    const localStorage = await chrome.storage.local.get();
-    new UnblockAllJobsManager(jobBoard, localStorage);
-    new RemoveHiddenJobsManager(jobBoard, localStorage);
-    new HiddenJobsListManager(jobBoard, localStorage);
-  }
-}
-
-export { JobBoardPopup };
